Split gocodemodel.ts into multiple files
Types are grouped by their relationships, making it easier to find
things instead of searching through a large, monolithic file.
