+++ packages/codegen.go/src/options.ts
@@ -5,7 +5,7 @@
 
 import { capitalize, comment } from '@azure-tools/codegen';
 import { values } from '@azure-tools/linq';
-import * as go from '../../codemodel.go/src/gocodemodel.js';
+import * as go from '../../codemodel.go/src/index.js';
 import { commentLength, contentPreamble } from './helpers.js';
 import { ImportManager } from './imports.js';
