+++ packages/codemodel.go/src/index.ts
@@ -1,0 +1,10 @@
+/*---------------------------------------------------------------------------------------------
+*  Copyright (c) Microsoft Corporation. All rights reserved.
+*  Licensed under the MIT License. See License.txt in the project root for license information.
+*--------------------------------------------------------------------------------------------*/
+
+export * from './client.js';
+export * from './package.js';
+export * from './param.js';
+export * from './result.js';
+export * from './type.js';
