+++ packages/typespec-go/src/tcgcadapter/adapter.ts
@@ -6,7 +6,7 @@
 import { clientAdapter } from './clients.js';
 import { typeAdapter } from './types.js';
 import { GoEmitterOptions } from '../lib.js';
-import * as go from '../../../codemodel.go/src/gocodemodel.js';
+import * as go from '../../../codemodel.go/src/index.js';
 import { packageNameFromOutputFolder, trimPackagePrefix } from '../../../naming.go/src/naming.js';
 import * as tcgc from '@azure-tools/typespec-client-generator-core';
 import { EmitContext } from '@typespec/compiler';
