+++ packages/typespec-go/src/tcgcadapter/types.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as naming from '../../../naming.go/src/naming.js';
-import * as go from '../../../codemodel.go/src/gocodemodel.js';
+import * as go from '../../../codemodel.go/src/index.js';
 import * as tcgc from '@azure-tools/typespec-client-generator-core';
 import * as tsp from '@typespec/compiler';
 import { values } from '@azure-tools/linq';
